CodeAlpha_Chatbot_for_FAQs/requirements.txt
crewai>=0.11.0
nltk>=3.8.1
spacy>=3.7.2
python-dotenv>=1.0.0
uv>=0.1.0
langchain>=0.1.0 